+++ src/VaultSharp/V1/AuthMethods/IAuthMethod.cs
@@ -4,6 +4,7 @@ using VaultSharp.V1.AuthMethods.AWS;
 using VaultSharp.V1.AuthMethods.Azure;
 using VaultSharp.V1.AuthMethods.Cert;
 using VaultSharp.V1.AuthMethods.GitHub;
+using VaultSharp.V1.AuthMethods.Kerberos;
 using VaultSharp.V1.AuthMethods.Kubernetes;
 using VaultSharp.V1.AuthMethods.LDAP;
 using VaultSharp.V1.AuthMethods.Okta;
@@ -58,6 +59,12 @@ namespace VaultSharp.V1.AuthMethods
         /// 
         /// </summary>
         ILDAPAuthMethod LDAP { get; }
+
+
+        /// <summary>
+        /// 
+        /// </summary>
+        IKerberosAuthMethod Kerberos { get; }
 
 
         /// <summary>